Move auxiliary functions and types to separate files
