+++ src/core/errors.py
@@ -1,0 +1,7 @@
+from enum import IntEnum
+
+
+class DatasetError(IntEnum):
+    NOT_FOUND = 111
+    NO_ACCESS = 112
+    NO_DATA_FILE = 113
